+++ apps/web/src/shared/ui/index.ts
@@ -1,1 +1,3 @@
 export { AuthCard } from "./auth-card";
+export { DatePicker } from "./date-picker";
+export { TimePicker } from "./time-picker";
